+++ django_server/fpf_sensor_service/models/__init__.py
@@ -1,0 +1,2 @@
+from .sensor_config import SensorConfig
+from .sensor_measurement import SensorMeasurement
