+++ django_server/fpf_sensor_service/models/sensor_measurement.py
@@ -1,13 +1,6 @@
-from django.db import models
 import uuid
-
-
-class SensorConfig(models.Model):
-    """
-    SensorConfig model with the id as a GUID and the intervall as an Integer
-    """
-    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-    intervalSeconds = models.IntegerField()
+from django.db import models
+from .sensor_config import SensorConfig
 
 
 class SensorMeasurement(models.Model):
