+++ django_server/fpf_sensor_service/views/__init__.py
@@ -1,0 +1,1 @@
+from .sensor_config_views import post_sensor_interval
